+++ sdk/communication/azure-communication-jobrouter/src/main/java/com/azure/communication/jobrouter/models/RouterJobNote.java
@@ -4,11 +4,21 @@
 package com.azure.communication.jobrouter.models;
 
 import java.time.OffsetDateTime;
+import java.util.Objects;
 
 /**
- * Note for closing job.
+ * Note for a job.
  */
 public class RouterJobNote {
+    /**
+     * Creates an instance of RouterJobNote class.
+     *
+     * @param message The message for the note.
+     */
+    public RouterJobNote(String message) {
+        this.message = Objects.requireNonNull(message, "'message' cannot be null.");
+    }
+
     /**
      * The time at which the note was added in UTC.
      */
@@ -17,7 +27,7 @@ public class RouterJobNote {
     /**
      * Message for the note.
      */
-    private String message;
+    private final String message;
 
     /**
      * setter for time.
@@ -35,16 +45,6 @@ public class RouterJobNote {
      */
     public OffsetDateTime getAddedAt() {
         return addedAt;
-    }
-
-    /**
-     * Setter for message.
-     * @param message Message for the note.
-     * @return this
-     */
-    public RouterJobNote setMessage(String message) {
-        this.message = message;
-        return this;
     }
 
     /**