+++ sdk/dataprotection/azure-resourcemanager-dataprotection/src/main/java/com/azure/resourcemanager/dataprotection/models/ResourceGuardProxyBaseResource.java
@@ -64,11 +64,13 @@ public interface ResourceGuardProxyBaseResource {
     interface Definition
         extends DefinitionStages.Blank, DefinitionStages.WithParentResource, DefinitionStages.WithCreate {
     }
+
     /** The ResourceGuardProxyBaseResource definition stages. */
     interface DefinitionStages {
         /** The first stage of the ResourceGuardProxyBaseResource definition. */
         interface Blank extends WithParentResource {
         }
+
         /** The stage of the ResourceGuardProxyBaseResource definition allowing to specify parent resource. */
         interface WithParentResource {
             /**
@@ -80,6 +82,7 @@ public interface ResourceGuardProxyBaseResource {
              */
             WithCreate withExistingBackupVault(String resourceGroupName, String vaultName);
         }
+
         /**
          * The stage of the ResourceGuardProxyBaseResource definition which contains all the minimum required properties
          * for the resource to be created, but also allows for any other optional properties to be specified.
@@ -100,6 +103,7 @@ public interface ResourceGuardProxyBaseResource {
              */
             ResourceGuardProxyBaseResource create(Context context);
         }
+
         /** The stage of the ResourceGuardProxyBaseResource definition allowing to specify properties. */
         interface WithProperties {
             /**
@@ -111,6 +115,7 @@ public interface ResourceGuardProxyBaseResource {
             WithCreate withProperties(ResourceGuardProxyBase properties);
         }
     }
+
     /**
      * Begins update for the ResourceGuardProxyBaseResource resource.
      *
@@ -135,6 +140,7 @@ public interface ResourceGuardProxyBaseResource {
          */
         ResourceGuardProxyBaseResource apply(Context context);
     }
+
     /** The ResourceGuardProxyBaseResource update stages. */
     interface UpdateStages {
         /** The stage of the ResourceGuardProxyBaseResource update allowing to specify properties. */
@@ -148,6 +154,7 @@ public interface ResourceGuardProxyBaseResource {
             Update withProperties(ResourceGuardProxyBase properties);
         }
     }
+
     /**
      * Refreshes the resource to sync with Azure.
      *