+++ sdk/resourcemanager/azure-resourcemanager-storage/src/main/java/com/azure/resourcemanager/storage/models/ServiceSpecification.java
@@ -17,6 +17,10 @@ public final class ServiceSpecification {
     @JsonProperty(value = "metricSpecifications")
     private List<MetricSpecification> metricSpecifications;
 
+    /** Creates an instance of ServiceSpecification class. */
+    public ServiceSpecification() {
+    }
+
     /**
      * Get the metricSpecifications property: Metric specifications of operation.
      *