+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -3,12 +3,14 @@
 ## 1.10.0-beta.2 (Unreleased)
 
 ### Features Added
+- Added `BrowserCustomizationOptions` to `InteractiveBrowserCredentialBuilder` to allow for customization of the browser window.
 
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+- Renamed `enableSupportLogging` to `enableUnsafeSupportLogging`. This is a breaking change from 1.10.0-beta.1.
 
 ## 1.10.0-beta.1 (2023-07-24)